requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-csp==3.8
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
psycopg2-binary==2.9.10
Pillow==11.1.0
python-dotenv==1.0.1
bleach==6.2.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
daphne==4.1.2
python-magic==0.4.27
whitenoise==6.8.2
gunicorn==23.0.0
psutil==6.1.1
